# PricetoolAPI.Application/DTOs/ProductDTO.cs
namespace PricetoolAPI.Application.DTOs
{
    public class ProductDTO
    {
        public int Id { get; set; }

        public string Name { get; set; }

        public string Type { get; set; }

        public decimal Principal { get; set; }



    }
}