# PricetoolAPI.Application/DTOs/SubsidiaryDTO.cs
namespace PricetoolAPI.Application.DTOs
{
    public class SubsidiaryDTO
    {
        public int Id { get; set; }
        public string Name { get; set; }
        public int Cvr { get; set; }
        public string Address { get; set; }

    }
}